Доработка сервиса wish

+++ src/modules/wishes/wishes.service.ts
@@ -1,19 +1,37 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { DataSource, Repository } from 'typeorm';
 
 import { User } from '@/modules/users/entities/user.entity';
 
+import { checkForbidden } from '@/common/utils/service/check-forbidden';
+import { checkHasEntity } from '@/common/utils/service/check-has-entity';
+import { TransactionService } from '@/common/utils/service/transaction';
+
+import { GET_WISH_DTO_ORM_OPTIONS } from './const/orm';
 import { CreateWishDto } from './dto/create-wish.dto';
 import { UpdateWishDto } from './dto/update-wish.dto';
 import { Wish } from './entities/wish.entity';
 
 @Injectable()
 export class WishesService {
+  private readonly transactionService: TransactionService;
+
   constructor(
+    private dataSource: DataSource,
     @InjectRepository(Wish)
     private readonly wishRepository: Repository<Wish>,
-  ) {}
+  ) {
+    this.transactionService = new TransactionService(dataSource);
+  }
+
+  async checkEditPermissions(user: User, id: Wish['id']) {
+    const wishData = await this.findOne(id);
+
+    const wish = checkHasEntity(wishData, 'WISH');
+
+    return checkForbidden(user, wish, wish.owner.id);
+  }
 
   async create(user: User, createWishDto: CreateWishDto) {
     const wish = this.wishRepository.create({
@@ -23,74 +41,85 @@ export class WishesService {
 
     await this.wishRepository.save(wish);
 
-    return this.wishRepository.findOne({
-      where: { id: wish.id },
-      relations: ['owner', 'offers'],
-    });
+    return this.findOne(wish.id);
   }
 
   async findLast() {
     return await this.wishRepository.find({
-      relations: ['owner', 'offers'],
+      ...GET_WISH_DTO_ORM_OPTIONS,
       order: {
         createdAt: 'DESC',
       },
       take: 40,
+      skip: 0,
     });
   }
 
   async findTop() {
     return await this.wishRepository.find({
-      relations: ['owner', 'offers'],
+      ...GET_WISH_DTO_ORM_OPTIONS,
       order: {
         copied: 'DESC',
       },
       take: 20,
+      skip: 0,
     });
   }
 
-  async findOne(id: number) {
-    return await this.wishRepository.findOne({
+  async findOne(id: Wish['id']) {
+    const wish = await this.wishRepository.findOne({
       where: { id },
-      relations: ['owner', 'offers'],
+      ...GET_WISH_DTO_ORM_OPTIONS,
     });
+
+    return checkHasEntity(wish, 'WISH');
   }
 
-  async update(id: number, updateWishDto: UpdateWishDto) {
+  async update(user: User, id: Wish['id'], updateWishDto: UpdateWishDto) {
+    await this.checkEditPermissions(user, id);
+
     await this.wishRepository.update(id, updateWishDto);
     return await this.wishRepository.findOne({
       where: { id },
-      relations: ['owner', 'offers'],
+      ...GET_WISH_DTO_ORM_OPTIONS,
     });
   }
 
-  async remove(id: number) {
-    const wish = await this.wishRepository.findOne({
-      where: { id },
-    });
+  async remove(user: User, id: Wish['id']) {
+    await this.checkEditPermissions(user, id);
 
     await this.wishRepository.delete(id);
-    return wish;
+    return;
   }
 
-  async copy(user: User, id: number) {
-    const wish = await this.wishRepository.findOne({
-      where: { id },
+  async copy(user: User, id: Wish['id']) {
+    const result = await this.transactionService.run<Wish>(async (manager) => {
+      const wishData = await manager.findOne(Wish, {
+        where: { id },
+      });
+
+      const wish = checkHasEntity(wishData, 'WISH');
+
+      const newWish: Wish = {
+        ...wish,
+        id: 0,
+        copied: 0,
+        owner: user,
+      };
+
+      const createdWish = manager.create(Wish, newWish);
+
+      await manager.save(createdWish);
+      await manager.increment(Wish, { id }, 'copied', 1);
+
+      const foundWish = await manager.findOne(Wish, {
+        where: { id: createdWish.id },
+        ...GET_WISH_DTO_ORM_OPTIONS,
+      });
+
+      return checkHasEntity(foundWish, 'WISH');
     });
 
-    if (!wish) {
-      throw new NotFoundException('Подарок не найден');
-    }
-
-    const newWish = {
-      ...wish,
-      copied: 0,
-      id: undefined,
-    };
-
-    await this.create(user, newWish);
-    await this.wishRepository.increment({ id }, 'copied', 1);
-
-    return newWish;
+    return result;
   }
 }
